requirements.txt
# LightRain Trading System - Python Dependencies

# Core Data & Analysis
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Database
psycopg2-binary==2.9.9

# Telegram Integration
python-telegram-bot==20.7

# Utilities
pytz==2023.3
logzero==1.7.0
python-dotenv==1.0.0

# Angel One Broker API
smartapi-python==1.3.5
pyotp==2.9.0

# Optional: For advanced features
# requests==2.31.0
# beautifulsoup4==4.12.2
